requirements.txt
Django==3.2.9
django-environ==0.8.1
psycopg2-binary==2.8.6
gunicorn==20.1.0